core: Move line wrapping code to another file
